+++ piecash/kvp.py
@@ -65,7 +65,7 @@ class SlotType(types.TypeDecorator):
 
 class DictWrapper(object):
     def __contains__(self, key):
-        for sl in self.slot_collection:
+        for sl in self.slots:
             if sl.name == key:
                 return True
         else:
@@ -76,7 +76,7 @@ class DictWrapper(object):
         assert not isinstance(key, int), "You are accessing slots with an integer (={}) while a string is expected".format(key)
         keys = key.split("/", 1)
         key = keys[0]
-        for sl in self.slot_collection:
+        for sl in self.slots:
             if sl.name == key:
                 break
         else:
@@ -89,7 +89,7 @@ class DictWrapper(object):
     def __setitem__(self, key, value):
         keys = key.split("/", 1)
         key = keys[0]
-        for sl in self.slot_collection:
+        for sl in self.slots:
             if sl.name == key:
                 break
         else:
@@ -100,9 +100,9 @@ class DictWrapper(object):
                 else:
                     sf = SlotFrame(name= key)
                 sf[keys[1]] = value
-                self.slot_collection.append(sf)
+                self.slots.append(sf)
             else:
-                self.slot_collection.append(slot(parent=self, name=key, value=value))
+                self.slots.append(slot(parent=self, name=key, value=value))
 
             return
         if len(keys)>1:
@@ -117,10 +117,10 @@ class DictWrapper(object):
     def __delitem__(self, key):
         if isinstance(key,slice):
             # delete all
-            del self.slot_collection[key]
+            del self.slots[key]
             return
         keys = key.split("/", 1)
-        for i, sl in enumerate(self.slot_collection):
+        for i, sl in enumerate(self.slots):
             if sl.name == keys[0]:
                 break
         else:
@@ -128,11 +128,11 @@ class DictWrapper(object):
         if len(keys)>1:
             del sl[keys[1]]
         else:
-            del self.slot_collection[i]
+            del self.slots[i]
 
 
     def iteritems(self):
-        for sl in self.slot_collection:
+        for sl in self.slots:
             yield sl.name, sl
 
     def get(self, key, default=None):
@@ -246,7 +246,7 @@ class SlotFrame(DictWrapper, Slot):
 
     guid_val = Column('guid_val', VARCHAR(length=32))
 
-    slot_collection = relation('Slot',
+    slots = relation('Slot',
                                primaryjoin=foreign(Slot.obj_guid) == guid_val,
                                cascade='all, delete-orphan',
                                collection_class=CallableList,
@@ -257,11 +257,11 @@ class SlotFrame(DictWrapper, Slot):
     @property
     def value(self):
         # convert to dict
-        return {sl.name: sl.value for sl in self.slot_collection}
+        return {sl.name: sl.value for sl in self.slots}
 
     @value.setter
     def value(self, value):
-        self.slot_collection = [slot(parent=self, name=k, value=v) for k, v in value.items()]
+        self.slots = [slot(parent=self, name=k, value=v) for k, v in value.items()]
 
 
     def __init__(self, **kwargs):
@@ -269,7 +269,7 @@ class SlotFrame(DictWrapper, Slot):
         super(SlotFrame, self).__init__(**kwargs)
 
 
-@event.listens_for(SlotFrame.slot_collection, 'remove')
+@event.listens_for(SlotFrame.slots, 'remove')
 def remove_slot(target, value, initiator):
     s = object_session(value)
     if value in s.new: