+++ tests/test_integration.py
@@ -95,7 +95,7 @@ class TestIntegration_EmptyBook(object):
 
         b["a/b/c/d/f"] = "2"
         session.sa_session.flush()
-        assert len(b["a"]["b"]["c"]["d"].slot_collection)==2
+        assert len(b["a"]["b"]["c"]["d"].slots)==2
 
         b["a/b/c/d/f"] = "5"
         assert b["a"]["b/c"]["d"]["f"].value == "5"
@@ -123,15 +123,15 @@ class TestIntegration_EmptyBook(object):
         assert {n for (n,) in session.sa_session.query(Slot._name)} == {'a' ,'a/b'}
 
         session.sa_session.flush()
-        assert len(b["a"].slot_collection)==1
-        assert len(b["a/b"].slot_collection)==0
+        assert len(b["a"].slots)==1
+        assert len(b["a/b"].slots)==0
 
         with pytest.raises(KeyError):
             b["a/b/c"]
 
         del b["a"]["b"]
         session.sa_session.flush()
-        assert len(b["a"].slot_collection)==0
+        assert len(b["a"].slots)==0
 
         with pytest.raises(TypeError):
             b["a"] = b